# Speed up inlist for strings and primitives