# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #814

 # Rationale for this change
<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

This gives a ~10-15% speed up for TCP-H query 12 (~24ms vs 28ms in memory) which spends quite some time in  the code to execute `l_shipmode in ('MAIL', 'SHIP')`

For the micro bench on primitives (f32 values):

```
filter_scalar in list   time:   [1.1045 ms 1.1285 ms 1.1607 ms]                                   
                        change: [-23.382% -21.418% -19.513%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  3 (3.00%) high severe
```

A large part comes from the use of the slower `BooleanArray::from_iter` and some more branching inside the loop.

This PR basically optimizes the easier cases when the list to compare to doesn't contain nulls.




# What changes are included in this PR?
<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

* Reusing the `compare_op_scalar` from arrow.
* Special casing on not-contains null and negated or not.

# Are there any user-facing changes?
<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->
Some code now returns `DataFusionError::NotImplemented` error instead of runtime panic.